+++ Tests/Extends/same-parameter-name.build.ps1
@@ -1,0 +1,9 @@
+﻿
+param(
+	[ValidateScript({"..\..\Tasks\Extends\Base\Base.build.ps1"})]
+	$Extends
+	,
+	$Base1
+)
+
+task .